Sequence counter as translation unit
